requirements.txt
pyserial==3.5
websockets==11.0.3
aiohttp==3.9.1
aiohttp-cors==0.7.0
requests==2.31.0
